Client Update
Moved the search-by-program page to the main page of the app. Considering this will be the main way in which most users use the app, it makes sense to have this be the star of the show. I've kept the search-by-credential page but it's essentially a hidden route. Updated the navigation to reflect this change. Kept in comments in case we'd like to revert later, or have navigation later.
